Move terminalProcessManager to node
