+++ src/tsconfig.strictNullChecks.json
@@ -294,10 +294,10 @@
 		"./vs/workbench/contrib/terminal/common/terminalMenu.ts",
 		"./vs/workbench/contrib/terminal/common/terminalService.ts",
 		"./vs/workbench/contrib/terminal/electron-browser/terminalInstance.ts",
-		"./vs/workbench/contrib/terminal/electron-browser/terminalProcessManager.ts",
 		"./vs/workbench/contrib/terminal/node/terminal.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcess.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcessExtHostProxy.ts",
+		"./vs/workbench/contrib/terminal/node/terminalProcessManager.ts",
 		"./vs/workbench/contrib/terminal/node/windowsShellHelper.ts",
 		"./vs/workbench/contrib/terminal/test/electron-browser/terminalColorRegistry.test.ts",
 		"./vs/workbench/contrib/terminal/test/electron-browser/terminalCommandTracker.test.ts",
