+++ src/vs/workbench/contrib/terminal/node/terminalProcessManager.ts
